srv/secure_mailvault/keychain/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
python-dotenv>=1.0
cryptography>=42.0
argon2-cffi>=23.0
